Refactor eth-client
