+++ packages/nitro-node/src/node/engine/chainservice/utils/utils.ts
@@ -5,7 +5,7 @@ import { EthClient } from '@cerc-io/nitro-util';
 
 // connectToChain connects to the chain at the given url and returns a client and a transactor.
 export async function connectToChain(chainUrl: string, chainPK: Buffer): Promise<[EthClient, ethers.Signer]> {
-  const client = await EthClient.dial(chainUrl);
+  const client = EthClient.dial(chainUrl);
 
   const txSigner = new ethers.Wallet(chainPK, client.provider);
