+++ packages/nitro-util/src/index.ts
@@ -9,6 +9,7 @@ export * from './constants';
 export * from './deploy-contracts';
 export * from './context';
 export * from './error';
+export * from './miscs';
 
 export {
   INitroTypes, ExitFormat, DepositedEventObject, AllocationUpdatedEventObject, ConcludedEventObject,
